Rename plugin modules
