+++ xicam/plugins/ezplugin.py
