+++ xicam/plugins/fittablemodelplugin.py
