+++ xicam/plugins/guiplugin.py
