+++ xicam/plugins/processingplugin.py
