+++ xicam/plugins/settingsplugin.py
