+++ xicam/plugins/widgetplugin.py
